herobrine: Increase heap size to boot CRD3.0
As the Qualcomm reference boards are using the 64MB SPI chip, they
need to increase the heap size as well in order to boot.  Since
herobrine and hoglin are sharing the herobrine defconfig now, we have
to increase the heap size for both boards (without having to create
another defconfig).  I did a comparison with cbmem on total boot time
and the affect of increasing the heap size is pretty negligible on the
herobrine board.

BUG=b:206581077
BRANCH=None
TEST=Qualcomm tested that this change booted on CRD 3.0
     Google tested that this change boots on herobrine board

Signed-off-by: Ravi Kumar Bokka <rbokka@codeaurora.org>
Change-Id: Ifd04936a7014d4d7b39d3dd3de1d0249937122ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3448809
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Queue: Shelley Chen <shchen@chromium.org>
Tested-by: Shelley Chen <shchen@chromium.org>