drivers/flash: Check SPI Cycle In-Progress prior start HW Seq
As per EDS, SPI controller sets the HSFSTS.bit5 (SCIP) when software
sets the Flash Cycle Go (FGO) bit in the Hardware Sequencing Flash
Control register.

This bit remains set until the cycle completes on the SPI interface.
Hardware automatically sets and clears this bit so that software can
determine when read data is valid and/or when it is safe to begin
programming the next command.

Software must initiate the next SPI transaction when this bit is 0.

Added non-blocking mechanism with `5sec` timeout to report back error
if current SPI transaction is failing due to on-going SPI access.

BUG=b:215255210

Signed-off-by: Subrata Banik <subratabanik@google.com>
Change-Id: Iea916239073fc6db801890b4549fb1d04f291ad8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3455264
Reviewed-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Commit-Queue: Subrata Banik <subratabanik@chromium.org>
Tested-by: Subrata Banik <subratabanik@chromium.org>