ui: Remove locale_id and others from ui_context
The following fields are stored in both ui_context and ui_state:

- locale_id (locale)
- disable_timer (timer_disabled)
- error_code

These fields in ui_state are used by the drawing code (ui_display()) to
reduce unnecessary screen redrawing. On the other hand, the same fields
in ui_context are used as "static" variables, and hence should be
preserved when changing screens.

Therefore, remove these fields from ui_context. Each time the UI state
is changed (from either ui_screen_change() or ui_screen_back()), we copy
the values of these 3 fields to the new state.

Because the locale_id is removed from ui_context, we need to add 'id' in
the ui_locale struct in order to initialize the language selection
screen.

BUG=b:172339016, b:216588026
TEST=make unit-tests
TEST=emerge-corsola depthcharge
TEST=In developer UI on krabby,
     1. Navigated to firmware log
     2. Pressed TAB to go to debug info
     3. Selected "Page down"
     3. Changed language
     4. Selected "Back"
     5. Pressed ESC
     6. Pressed Ctrl+L to trigger error box
BRANCH=none

Change-Id: Ia263ca2b90b781390a39b190346567ee99f58132
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3459251
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>