mtk_mmc: fix device hang when writing data in HS200
The MediaTek MMC controller has the ability to detect the busy state of
the eMMC device. For example, it can wait for Busy on CMD12 after CMD25
and write programming timeout.
If we enable this function, WRDTOC is used for R1b Busy wait timeout.
But the timeout will be too short, and lead to a timeout interrupt.
In this patch we will turn off this function, so the busy state of the
eMMC device will be detected by sending cmd13 - which is already implemented.

BUG=b:219654107
BRANCH=cherry
TEST=boot to kernel ok and suspend/resume test pass

Signed-off-by: Wenbin Mei <wenbin.mei@mediatek.com>
Change-Id: I6010fcfedcae92297a539b7072f64d2f7c5f1812
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3473333
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>