ui: Skip reinit() when screen doesn't change
In ui_developer_mode_boot_external_action(), if there is no external
disk found, ui_screen_change() will be called to change to the developer
boot external screen, even if we are already in that screen. However,
because of the way ui_screen_change() is implemented, the reinit() hook
will be called in this case, which causes the selected item index to be
reset.

To avoid reinit() being called repeatedly in the developer boot external
screen, add a check at the beginning of ui_screen_change() to return
early if we are already in the requested screen.

BUG=b:222197956
TEST=make unit-tests -j
TEST=make screenshot
TEST=emerge-corsola depthcharge
TEST=UP/DOWN was functional in "developer boot external" screen
BRANCH=cherry,brya,guybrush

Change-Id: Iab633a1c3608d08dc57680b4db718faf8715124a
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3498944
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>