depthcharge: increase EC erase timeout to 30 seconds
Some ECs like the IT81302 used on Nereid can take a long time to erase;
the IT81302 in particular is rated for a typical erase time (assuming a
384k EC firmware image) of around 5 seconds, but a maximum of nearly 30.
Increase the timeout for erase commands to match that worst-case.

This shouldn't negatively impact platforms that are faster to erase,
because they will only wait that long if something has gone wrong in
which case this gives them more time to finish: timing out would require
the user to recover the system, whereas waiting longer improves the
chances of the EC eventually finishing the operation so the user doesn't
need to recover.

Because flash erase is the only command that is expected to return
IN_PROGRESS, this adds a message if any other command does so
unexpectedly and prints timing information on completion so devices can
be monitored to see if they are performing badly (because 30 seconds is
a pretty long time, and we don't expect devices to need that much time
in most situations).

BUG=b:222987250
TEST=Nereid (still) boots with EC software sync enabled
BRANCH=none

Change-Id: Id79f000e6c18ecc66d05bd66fd34536761ace96f
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3607309
Reviewed-by: Kangheui Won <khwon@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>