driver/bus/i2c: Add configurable I2C transfer timeout
This patch introduces CONFIG_I2C_TRANSFER_TIMEOUT_US,
which controls how long to wait for an I2C devices to
produce/accept all the data bytes in a single transfer.
(The device can delay transfer by stretching the clock of
the ack bit.)

The default value of this new setting is 500ms.  Existing
code had timeouts anywhere from tens of milliseconds to a
full second beween various drivers.  Drivers can still have
their own shorter timeouts for setup/communication with the
I2C host controller (as opposed to transactions with I2C
devices on the bus.)

In general, the timeout is not meant to be reached except in
situations where there is already serious problem with the
boot, and serves to make sure that some useful diagnostic
output is produced on the console.

This patch mimics a prior change to coreboot:
https://review.coreboot.org/c/coreboot/+/62278

BUG=b:230377468
TEST=Booted Brya board
BRANCH=brya

Change-Id: I7635c44ff20020a5b6f7375d0396adf3b04e65a8
Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3611663
Reviewed-by: Raul Rangel <rrangel@chromium.org>