vboot: Wipe the memory before any non-normal boot
We only requested the memory wipe for dev/rec boot before, this was
sufficient until the introduction of the diagnostic boot (MiniDiag).

After the diagnostic boot and the ctx->boot_mode are introduced, we
could consider change the wiping logic in vboot_check_wipe_memory().

Change it to 'wipe the memory unless it is a normal boot'.

BUG=b:231398382
BRANCH=none
TEST=make unit-tests

Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: I22d0d5f64a653baf98f67bfa0d304969220f672a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3627429
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>