ui: Always show common developer options
People seem to be confused of how to go from developer mode back to
secure mode (normal mode) because they don't see the "Return to secure
mode" button in the UI. This happens when the GBB flag
FORCE_DEV_SWITCH_ON is set, which is the default value of locally built
firmware.

To improve UX, we change the behavior to always showing the button,
regardless of the GBB flag. When it is not allowed to return to secure
mode, an error message will be shown to notify the user.

For consistency's sake, we also show the "Boot from external disk"
button unconditionally. For "Select alternate bootloader", however,
since it's rarely used, only showing it when necessary seems to be the
better option.

BUG=b:221372833
TEST=make unit-tests
TEST=make test-screenshot
TEST=emerge-corsola depthcharge
TEST=On cherry, set FORCE_DEV_SWITCH_ON and checked the menu items
TEST=On cherry, unset FORCE_DEV_BOOT_USB and checked the menu items
BRANCH=none

Change-Id: Iac8aec0fd5c40a16886188723fda3cc70635c33e
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3637127
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>