ui: Allow manual recovery even if no locale found
As a side effect of CL:3459251, when the locale file is not found in
CBFS, the device will fail to go to the UI loop, consequently
disallowing the user to do anything including USB recovery. To use best
efforts to allow manual recovery (and other developer UI actions), use
the hard-coded English locale as the last resort.

When a non-zero locale_id is not found, we'll print a warning message:

 Locale 123 not found, falling back to locale 0

and try to find locale 0. If it still cannot be found, there'll be
another warning printed:

 Fallback locale 0 still not found, assuming "en"

Also add a new test 'test_manual_ui_locale_not_found' to make sure the
recovery process is still functional in such an extreme case.

BUG=b:232191856
TEST=make unit-tests -j
TEST=emerge-cherry depthcharge
BRANCH=none

Change-Id: I9574e65a9b4133bd1955b6e45ed9fb99b2fcc263
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3642984
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>