vboot: Disable TPM in cleanup hook
TPM should be disabled before booting into miniOS. However, we need TPM
to generate a random number for kASLR, so we can't disable TPM before
that.

A new vboot context flag VB2_CONTEXT_DISABLE_TPM is added in CL:3653207.
When it is set, we disable TPM in the cleanup hook function. This avoids
the "TPM failed to populate kASLR seed buffer." error when doing
internet recovery.

BUG=b:223662000, b:232743820
TEST=make unit-tests -j
TEST=emerge-cherry depthcharge
TEST=Booted into miniOS in less than 2 seconds on tomato
BRANCH=cherry

Cq-Depend: chromium:3653207
Change-Id: I148702840f161cd8e2da1bb3da25fef3448b7cac
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3653659
Reviewed-by: Julius Werner <jwerner@chromium.org>