video: Clean up display for all exit conditions
We currently only clean up the display (stopping the display controller
output and turning off the backlight) on kernel handoff. The assumption
is that shutdown/reboot will "automatically" turn off the display
anyway, but power sequencing can be tricky, especially for panels. In
one Pazquel (trogdor) case, we see a bunch of colored stripes flicker
across the display if we just turn off the machine while it's running.

This patch changes the display cleanup handler to run for all kinds of
exits from depthcharge, just to be on the safe side.

BRANCH=trogdor
BUG=b:233715471
TEST=QC confirms this fixed Pazquel flicker issue.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I23c40162df04e03cca8dfe3ef4752875b55a0812
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3668538
Reviewed-by: Shelley Chen <shchen@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>