diag: Fix buf out-of-range issue
During diagnostic steps, we need to append several texts to a buffer by
using snprintf() multiple times.

However, if a smaller buffer is assigned, there will be fatal memory
accessing issues since we misused the return value of snprintf().

This did not cause any error because the buffer was fixed-size and very
large, but we need to fix this for further refactor works.

Change the buffer size in diagnostics_init to 1 since we only need to
know whether the test log (e.g. Storage self-test) is supported or not.

BRANCH=none
BUG=b:234682419
TEST=Test MiniDiag on brya4es
TEST=make clean-unit-tests && make unit-tests -j

Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: I8f98a39871b6c2ad1c74646c07dc30ee4d47c92a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4027590
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Tested-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>
Commit-Queue: Subrata Banik <subratabanik@chromium.org>