storage/mmc: get access mode from coreboot
When mtk_emmc_early_init is enabled, some eMMC devices may enter
the ready state in coreboot and the MMC driver will not send CMD1
to get the access mode. Then the MMC driver will assume that the
device is in the byte mode and works in the wrong mode. So the MMC
driver needs to get the access mode from coreboot.

BUG=b:234672726
BRANCH=cherry
TEST=boot ok

Signed-off-by: Wenbin Mei <wenbin.mei@mediatek.com>
Change-Id: Ide81af2c9f3b4d3bb663d7885682b8d456b307ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3688903
Reviewed-by: Yidi Lin <yidilin@chromium.org>
Reviewed-by: Wenbin Mei <wenbin.mei@mediatek.corp-partner.google.com>
Tested-by: Wenbin Mei <wenbin.mei@mediatek.corp-partner.google.com>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>