tests: Makefile: Fix relative paths for upstream coreboot build
When building depthcharge in upstream coreboot, the path to which the
depthcharge repository gets checked out is not payloads/depthcharge,
it's payloads/external/depthcharge/depthcharge. Therefore, we'll need
two more .. in the path to get back to the coreboot root. (Also, the
coreboot root does not contain a directory called `coreboot`, it's just
the root.)

BRANCH=none
BUG=none
TEST=none

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: Ibb0d1a3066603e91fdd425a02b8e3451622dc988
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3698637
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>