drivers/storage/ufs: Use SCSI_FLAG_FUA for read function
A Forced Unit Access (FUA) bit set to zero specifies that the
device server may read the logical blocks from the volatile cache
(if any), the specified data pattern for that LBA (e.g., the data
pattern for unmapped data), the non-volatile cache, or the medium.

An FUA bit set to one specifies that the device server shall read
the logical blocks from the specified data pattern for that LBA,
the non-volatile cache (if any), or the medium.

To read the logical blocks from the medium, enable SCSI_FLAG_FUA
for read.

BRANCH=cherry
BUG=b:233327674
TEST=build & read correct data from UFS

Signed-off-by: Bayi Cheng <bayi.cheng@mediatek.corp-partner.google.com>
Change-Id: Ieb3b4543af4c9549b75189ecc6cf8e378b3c90d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3717117
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>