Shotzo: Add handling for recovery button
Shotzo has physical recovery button,
this CL is to configure it as physical presence button
thus we can read the recovery status from the cr50.

Because the deconfig dedede_alc1050_amp also uses the dedede board.c,
so it needs to add the same config into it.

BUG=b:235303242, b:235983675
BRANCH=firmware-dedede-13606.B
TEST=emerge-dedede depthcharge

Change-Id: I930f3fb0ca45503a9f7b60bc8dddf9c3d0a2eddb
Signed-off-by: Tony Huang <tony-huang@quanta.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3708308
Reviewed-by: Zhuohao Lee <zhuohao@chromium.org>