corsola: Delay 500ms after USB VBUS enabled
Even if a USB disk is inserted prior to AP boot, it cannot be detected
in the first few iterations of UI loop (in developer/recovery mode).
This can lead to confusing UI behavior. For example, if the user presses
Ctrl+U in developer screen before USB detection, the "Plug in your
external disk" screen will be shown for less than a second, and then the
device will automatically boot from the USB disk.

To make sure USB is detected before showing the firmware UI, delay 500ms
after USB VBUS is enabled. The extra delay is added only when USB
initialization is performed, so normal boot time won't be affected.

BUG=b:235552939
TEST=emerge-corsola depthcharge
TEST=Quickly pressed Ctrl+U in developer screen, and didn't see the
     "Plug in your external disk" screen
BRANCH=none

Change-Id: I5e5810141f48b1fcfada4282c672a57e780afd1b
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3713045
Reviewed-by: Hung-Te Lin <hungte@chromium.org>