tests: Add initial zero coverage data
Currently, when calculating the coverage percentage, only the files
built by unit tests are included. This leads to an overestimate of the
coverage, and the weird situation where adding a new unit test decreases
the overall coverage.

To include all the files that are built for depthcharge.elf, generate
initial zero coverage data from a regular depthcharge build (for a
specific board ${BOARD}), and pass it to lcov for coverage calculation.

For the libpayload dependency, building it from source with ARCH_MOCK
doesn't work because of missing various headers needed for different
platforms. Therefore, we require passing the libpayload install
directory ${LIBPAYLOAD_DIR} to the coverage-report make target.

Usage:

$ COV=1 make unit-tests -j
$ COV=1 make clean-coverage-report
$ LIBPAYLOAD_DIR=/build/corsola/firmware/kingler/libpayload/libpayload \
  BOARD=corsola COV=1 make coverage-report

where ${BOARD} must be compatible with the platform ${LIBPAYLOAD_DIR} is
built with.

BUG=none
TEST=COV=1 make unit-tests -j &&
     LIBPAYLOAD_DIR=/build/corsola/firmware/kingler/libpayload/libpayload \
     BOARD=corsola COV=1 make coverage-report
BRANCH=none

Cq-Depend: chromium:3764618
Change-Id: I6a577e5b85e07aba7feb80c75208fe178fcf5fa1
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3759282
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>