Sabrina: Remove embargoed name
'Mendocino' was an embargoed name and could previously not be used
in references to Skyrim.  Coreboot has references to sabrina both
in directory structure and in files. This will make life difficult
for people looking for Mendocino support in the long term. The code
name should be replaced with "mendocino".

BUG=b:239072117
TEST=Builds
BRANCH=none

Change-Id: Ia1918c934e557124b85495d1ae2fc719d0c00959
Signed-off-by: Jon Murphy <jpmurphy@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3764023
Reviewed-by: Tim Van Patten <timvp@google.com>
Commit-Queue: Tim Van Patten <timvp@google.com>
Reviewed-by: Robert Zieba <robertzieba@google.com>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>