drivers/storage/info: Always compile all types of members in the unions
Always compile all types of members in the unions: HealthInfo.data and
StorageTestLog.data.

The current generic storage health information union was firstly
introduced in CL:2193318. It leveraged preprocessors to save spaces.
However,

(1) The NvmeSmartLogData is 512 bytes, the MmcHealthData is 36 bytes,
and the NvmeTestLogData is 22 bytes. All the members are quite small
compared to the modern DRAM size. (at least 4GB)

(2) The current implementation, `#ifdef CONFIG_XXX`, is actually always
true. (The correct usage should be: `#if CONFIG(XXX)`) So we could
ensure that there is no potential size issue currently.

(3) Hiding some members in the compile time is tricky. We'd have to take
good care of the dependencies to prevent `union has no member` and
`defined but not used` compile errors.

Hence, we decide to remove these include guards and always compile all
members in these unions.

BUG=b:156998867
BRANCH=none
TEST=emerge-brya depthcharge (A board with NVMe but no eMMC)

Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: Ide524cf1822a7ceca3c29bccd2ab9b453c5c5891
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3784573
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>