fit: Simplify kernel randomness mixing
We're currently trying to stretch 64 bytes of TPM-provided randomness to
72 bytes of kASLR seed (8 bytes) and RNG seed (64 bytes) by appending
its hash behind it. This is actually not necessary (and just potentially
confusing/inaccurate to the kernel's randomness bookkeeping) because
there's no requirement that the RNG seed must exactly be 64 bytes. The
kernel will mix in exactly as much as given, and the kernel's CRNG will
consider itself "initialized" as soon as it has at least 32 bytes. So
let's just give it 56 bytes and remove the need to calculate an extra
hash here.

BRANCH=none
BUG=b:240624460
TEST=Booted CoachZ.

Signed-off-by: Julius Werner <jwerner@chromium.org>
Change-Id: I5eb358ab89a9322b5f3f13e4e763f9de9d0af8f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3831331
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>