util/kconfig_lint.pl: Adapt to depthcharge
The source code of "net" doesn't follow depthcharge's coding style, and
contains lots of macro definitions that share the same names with
Kconfig options. Therefore, instead of fixing the code, it's easier to
just ignore the whole "net" directories.

Also exclude src/Makefile.inc, which contains an explanation of the
CONFIG(option) macro.

Avoid searching configs within payloads' Makefiles, which are
specifically for the coreboot repository.

Ignore the LP_* options, which are defined in libpayload.

BUG=b:239779216
TEST=./util/kconfig_lint.pl -w
BRANCH=none

Change-Id: Ib50e6341ef6f15704cf8fc8f1d34c2ffaa2ea0da
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3870365
Reviewed-by: Yidi Lin <yidilin@google.com>
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>