shotzo: add a link to the dedede board setting
We merge the shotzo to the dedede OS image and the shotzo will be one
of unibuild projects under the dedede. To support the different depthcharge
settings like physical_presence_recovery, we need to have a dedicated
board setting for the shotzo. As a result, we add a link for the shotzo
to link with the dedede board setting.

BUG=b:235983675
BRANCH=firmware-dedede
TEST=1.build pass
     2. flash the firmware to the device and check the recovery mode

Change-Id: I7c2262f7dafdabbdc349ca26c3c662653b99c378
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3895338
Reviewed-by: Derek Huang <derekhuang@google.com>
Commit-Queue: Derek Huang <derekhuang@google.com>