shotzo: add the new board setting
We merge the shotzo to the dedede OS image and the shotzo will be one
of unibuild projects under the dedede. To support the different depthcharge
settings like physical_presence_recovery, we need to have a dedicated
board setting for the shotzo. As a result, we clone the dedede setting
for the shotzo.

BUG=b:235983675
BRANCH=firmware-dedede
TEST=1.build pass
     2. flash the firmware to the device and check the recovery mode

Change-Id: I81401d44517b2d161d356354bdb63e0deb77b7b5
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3898010
Reviewed-by: Liam Flaherty <liamflaherty@chromium.org>
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-by: Derek Huang <derekhuang@google.com>