vboot/ui/navigation: Add an exit() hook for screens
Some screens need to take extra actions while exiting from them. The
actual scenarios are:
1. Diagnostics test items want to proactively flush the cache on exiting
   a test item. This is introduced at CL:3826077 with different
   workarounds.
2. Diagnostics test items want to mark current test as "ERROR" if
   exiting the item abnormally.
3. Diagnostics root screen wants to generate a summary after powering
   off.

In CL:3826077, we leveraged the reinit() hooks in the root screen as a
workaround. But with more and more use cases appearing, we would want to
introduce the permanent solution of intruducing the exit() hook.

Also add the API of ui_screen_cleanup(). This should be called once
while the ui_loop is finished.

BUG=none
BRANCH=none
TEST=emerge-cherry depthcharge
TEST=make clean-screenshot && make test-screenshot -j
TEST=make clean-unit-tests && make unit-tests -j

Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>
Change-Id: Ib99bafee202434b38bed9975c78a4514514964be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/3924008
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>