herobrine: Do not let devices boot from eMMC if NVMe sku
Currently, we allow devices to boot from both eMMC and NVMe for the
NVMe sku.  Additionally, eMMC is prioritized, but that is incorrect
behavior.  In fact, we do not need to allow for booting from eMMC at
all for NVMe skus, so removing this initialization from depthcharge.

BUG=b:271535432
BRANCH=None
TEST=On herobrine board, make sure can still boot from eMMC
     Verified on Zombie will try to boot from NVMe.

Change-Id: I9e31a12125f64c2b554627e342af445f55e5813a
Signed-off-by: Shelley Chen <shchen@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4309819
Reviewed-by: Douglas Anderson <dianders@chromium.org>
Tested-by: Gavin Lee <gavin.lee@ecs.corp-partner.google.com>
Reviewed-by: Gavin Lee <gavin.lee@ecs.corp-partner.google.com>
Commit-Queue: Shelley Chen <shchen@chromium.org>