dibbi: Increase heap size
The firmware screens are missing text when using a Type-C external
display, due to an allocation failure when loading the firmware screen
files from CBFS. Increase the heap size to avoid running out of memory.
This is the same as how CL:2491381 fixed the same issue on volteer.

BRANCH=dedede
BUG=b:283196347
TEST=Firmware screens display properly over Type-C

Change-Id: Ic05e46fe5327cf36c175de68e63288514b4ae851
Signed-off-by: Reka Norman <rekanorman@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4547634
Tested-by: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>