Makefile: Use patsubst to support sub folders named as 'src'
The 'subst' will replace all occurrences of the pattern. So if a sub
folder in source tree has a directory named 'src' that will be also
incorrectly replaced.

For example, if we have 'src/ui/src/a.c', then in the substituted text
it will be 'ui/a.c'.

To fix that we should use patsubst when replacing the 'src' pattern.

BUG=None
TEST=emerge-$BOARD depthcharge
BRANCH=None

Change-Id: Ibc6af30fafaccc203c995ddf81697fcb8290080c
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4784467
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>