netboot: Drop CLI feature from netboot image
Previously, the CLI feature was enabled for both depthcharge.elf
and dev.elf. This meant that all types of images (normal, serial,
and netboot) could launch into the CLI if the CONFIG_CLI configuration
was enabled.

This patch limits the CLI debug feature to the serial image alone.
This is because the serial image is the only image that includes
dev.elf.

BUG=b:268448199
TEST=Able to launch CLI only with serial image.

Change-Id: I349ca8c05a0a73bb644373934f0d10235ff11eab
Signed-off-by: Subrata Banik <subratabanik@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4834895
Commit-Queue: Julius Werner <jwerner@chromium.org>
Tested-by: Subrata Banik <subratabanik@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Auto-Submit: Subrata Banik <subratabanik@chromium.org>