ui: Add support for headless devices
Currently the screen argument of DisplayOps.display_screen() is still of
the old deprecated type VbScreenType_t. To support headless devices in
the new firmware UI, migrate all the existing implementations of
display_screen() to the new screen type ui_screen.

One difference of VbScreenType_t and ui_screen is that, the latter
intentionally doesn't define the blank screen. In the legacy UI,
VB_SCREEN_BLANK was used to clear the screen before jumping to kernel.
However now that is done by the DisplayOps.stop() hook. Therefore, also
add stop() to those drivers.

Although the display of headless devices is through LED, the defined
screens and menu items are still almost the same as Chromebooks. It
seems unnecessarily complicated for the UI code to add lots of
conditionals of CONFIG(HEADLESS), just in order to drop all unsupported
screens for headless devices. Therefore, remove the only existing
conditional from vboot/ui/screens.c. If we want to properly support
headless devices in the future, a separate set of screens should be
defined for them.

BUG=b:151200757
TEST=make unit-tests -j
TEST=make test-screenshot
BRANCH=none

Change-Id: I28dd75f5f039e9162d4fc8c603b4044fc38925a6
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4867283
Reviewed-by: Hung-Te Lin <hungte@chromium.org>