payload: Correct decoding of simple ELF
The header is big endian, so the type needs to be converted just like
the other fields.

This bug was fixed as part of a CBFS migration in
https://crrev.com/c/3461453 (see patch 2)

Bring this fix in, so that altfw works as expected.

BUG=b:303125519
BRANCH=brya
TEST=emerge-brya depthcharge chromeos-bootimage
See that altfw boots now

Change-Id: I5b71b36a76470ff071aaf9638a10b25eda38b5a0
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4916811
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Jakub Czapiga <czapiga@google.com>
Reviewed-by: Jakub Czapiga <czapiga@google.com>