drivers/ec/cros: Disable lid close EC event if has_external_display
With CL:4845848, when an external display is attached and enabled, lid
close in firmware screens is expected to NOT shutdown the device. That
CL modified the shutdown condition in check_shutdown_request(). The
problem is, on Intel devices, ui_is_lid_open() actually always returns
the same value, which comes from "lid" GPIO in lib_sysinfo.gpios.
Therefore, only when we boot the device with lid closed,
ui_is_lid_open() will return 0 in depthcharge.

The situation of closing lid in firmware screens is handled differently
on Intel devices, which is actually done by an SMI handler defined in
coreboot. As a result, when performing the following steps:

1. Attach an external display
2. (In dev mode) Power on the device with lid closed
3. In developer screens, open the lid
4. Still in developer screens, close the lid

the device will shutdown in step 4.

To fix the problem, add the same condition of has_external_display() to
cros_ec_disable_lid_shutdown_at_startup().

BUG=b:299137940
TEST=make unit-tests -j
TEST=emerge-brya depthcharge
BRANCH=none

Change-Id: Id9347f5698ce177fd66b2ae8b7286bc642e3c5fe
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4935503
Reviewed-by: Subrata Banik <subratabanik@chromium.org>