ps8805: Handle when the revision register contains 0x4
On my homestar, I did this:
* Flashed an old firmware (13577.702.0) on the device and booted. Let
  it update the TCPC chip.
* Flahshed a new firmware (13577.882.0) on the device and then ran a
  test that would constantly do an EC "reboot" randomly every 4 to 33
  seconds. The 4 seconds is because it takes about 4 seconds for the
  TCPC firmware to start and 33 because it takes about 29 seconds for
  the TCPC firmware to finish.

After a while, the homestar was unable to recognize either TCPC chip,
was unable to use USB, and was unable to charge. :(

Digging, I found that _both_ the EC and depthcharge were confused
because the "revision" register (page 0, register 0x62) contained
0x44. Both depthcharge and the EC look at this register and mask it
with 0xf0 to determine between the A2 stepping and the A3 stepping. If
depthcharge sees 0xa0 it assumes A3. Otherwise it assumes A2. If the
EC sees 0xa0 it assumes A3, if it sees 0x00 it assumes A2, and for
anything else it is an error.

Digging more, I found that I could get a Lazor (which has the A2
stepping) into the exact same state, which means that when we're in
this state we can't tell the difference between PS8805_A2 and
PS8805_A3.

According to Parade we can put A3 firmware on a PS8805_A2 and it's not
a problem, so if we run into this situation then we'll just assume A3.

Interestingly enough, once you boot a PS8805_A2 with A3 firmware then
after reboot it properly identifies itself as an A3. This means that
on the next boot we'll actually switch back, so we'll really only be
using the A3 firmware for a very short period of time.

BUG=b:304772621
TEST=My broken device recovered itself

Change-Id: Ib6247e6a42e0bf4775ad18b8a8e2dd5e2821f615
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4939349
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>
Tested-by: Wai-Hong Tam <waihong@google.com>