coreboot: expose FW config in device tree
Similar to how we've plumbed SKU ID, board ID and RAM code.

This should provide the OS with more information on how to do per-board
device probing, should we choose to do that.

Note that the Kconfig options FW_CONFIG and FW_CONFIG_SOURCE_CHROMEEC_CBI
need to be enabled in coreboot for it to read and export FW config values
in the coreboot tables.

BRANCH=corsola
BUG=b:157692450
TEST=boot steelix; check /proc/device-tree/firmware/coreboot/fw-config

Change-Id: I279a849f6084de246e79e237c2e89d1e1b87a5ef
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/4975217
Reviewed-by: Yidi Lin <yidilin@chromium.org>