ui: Tell the EC about each firmware screen
Send a message to the EC with information about the firmware screen.
For now the menu items are not included, since they don't seem to have
stable values. It might be best for each item would have a globally
unique value no matter what screen it is in.

Add a comment pointing to the FAFT code which relies on the current
numbering.

Note: Future work will likely renumber the screens to fit into a byte,
since there are only a few dozen.

BUG=b:300525571
BRANCH=none
TEST=manually tested on skyrim in dev mode with updated EC and see a
message:
   23-11-10 10:11:01.194 AP_FW 300

Cq-Depend: chromium:5014199
Change-Id: Ib0ee45bf1c2245900b2f920cfd438bc9202cba93
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5020949
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>