tests/vboot: Use the real log.c for UI tests
The UI integration tests tests/vboot/ui-*-test.c should run as much
production code as possible, including the log processing code in log.c.
However currently we are using mock ui_log_init() in these tests.

Remove the mock log.c and use the real log.c for UI tests. For the debug
info screen, since we already have a working vboot workbuf which
contains all the required information for the debug info, the debug info
is created in the exact same way as the production code. For the
firmware log, however, we use a mock cbmem_console_snapshot() as before.
The broken UI test has a special cbmem_console_snapshot() mock for the
log screen test cases. The other UIs share the other common mock
cbmem_console_snapshot() defined in tests/mocks/cbmem_console.c, which
is simplified in this patch.

The mock ui_get_log_textbox_dimensions() is moved to the mock layout.c
because that's where it's supposed to be in the first place. A helper
macro SET_LOG_DIMENSIONS() is created to replace
WILL_CALL_UI_LOG_INIT_ALWAYS() and friends.

The log screen test cases are reogranized in ui-broken-test.c:

1. test_debug_info_one_page() is merged into test_debug_info().
2. test_firmware_log() is extended to exercise the keyboard shortcuts in
   the firmware log screen. test_debug_info_three_pages() is removed.

BUG=b:163301827
TEST=make unit-tests -j
BRANCH=none

Change-Id: I33de924395b0b8d78d251be063ddde6ee5d1fd3b
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/depthcharge/+/5043272
Reviewed-by: Hsuan Ting Chen <roccochen@chromium.org>